Streamline workflow, test Akaso fixes.
generate-ffmpeg.py:

Renamed functions in an effort to make it easier to read the code.

Renamed output of concat filter to concatinated.mp4 to make things easier to follow.

Prefix temporary files with prefix "filelist-" and put them in the individual camera
directories.  This allows them to stick aroud in case things need to be re-run.

Align the camera files by trimming the concatinated files.  Trimming is a fast operation
and elminates the need for all of the SETPTS operations that cause so many issues later.
Output goes in combined.mp4.

Remove scale= filter and setpts filter, replace with fps=fps=30 on input.
Continue to tile with xstack.

The 4 camera audio streams are all mono.  Cameras 1 and 3 have microphones pointed
away from the action.  Use only the audio from cameras 2 and 4, and make camera2
be the left channel and camera 4 be the right channel for stereo output.

Change metadata to output one per line to make them easier to read in the output.

Remove the final video trim as it is no longer required with the trim being done
per-camera up front.

README.md, camera-quirks.md

Update on Akaso's initial feedback and some testing to date to work
around the duplicated and dropped frame issue.