Elminiate warnings about audio.
The audio filter was taking the two mono streams from the camreas
and creating two stero streams, one panned hard left, the other hard
right.  A merge then tried to merge thes into a stereo stream
generating:

[Parsed_amerge_8 @ 0x7fac48c0e880] Input channel layouts overlap: output layout will be determined by the number of distinct input channels
[Parsed_pan_6 @ 0x7fac1910dc40] Pure channel mapping detected: 0 M
[Parsed_pan_7 @ 0x7fac48c0df80] Pure channel mapping detected: M 0

The correct answer is not to pan, give the two mono streams directly
to amerge and the first will become the left channel and the second
will become the right chnnnel.